Significantly improve log file format
This commit changes the log file format from a binary dump of all read
data to a text file recording all transactions. This makes the log files
much more descriptive as well as easier to read: you don't have to use
xxd, the messages are chopped up nicely and all the context is there.

It also allows (and implements) significant improvement of the replay
functionality to support out-of-band replay. The replay no longer relies
on identical program flow to the replayed run. This makes no difference
while there is a single, statically defined, flow - however once
user-interactivity is added, this will be very powerful.