Add initial unit test
Not 100% about the layout. Don't love it. Trying to do it as Bazely as
possible, but Bazel is fighting me a bit. And I don't want to go down
the 'dump all your source, tests and headers in one dir' route that the
Google-maintained projects seem to follow to fix it.