Re-implement read using iterators
Much more maintainable than the weird way I was doing it, and it
makes representing an empty iterator much more straight forward. I
was trying desparately to avoid some magic 'is_empty' state before
and was struggling to see a way through. Could be made
bi-directional, but that's not actually desireable for the serial
interface.