Remove internal EngineParameter methods from API
It's desireable to keep the definition of these internal methods
alongside the definition of the other methods that handle
EngineParameters - and likewise with the tests. But they're
implementation details of ConsultInterface, so should not be exposed on
the API.

To achieve this separation, I've added a new internal-only header that
does not form part of the library's API. This also nicely removes the
common.h dependency, although I may well go back and 'rangify' the API,
so I'm not removing that from the API just yet.